chore: move selection and transform events to separate files
