+++ Scribble/Services/CanvasStateService/CanvasStateService.cs
@@ -7,6 +7,7 @@ using Scribble.Services.CanvasStateService.State;
 using Scribble.Services.MultiUserDrawing;
 using Scribble.Shared.Lib;
 using Scribble.Shared.Lib.CanvasElements;
+using Scribble.Shared.Lib.Events;
 using SkiaSharp;
 
 namespace Scribble.Services.CanvasStateService;
