+++ Scribble/Services/CanvasStateService/Handlers/SelectionReplayHandler.cs
@@ -4,6 +4,7 @@ using Scribble.Services.CanvasStateService.State;
 using Scribble.Shared.Lib;
 using Scribble.Shared.Lib.CanvasElements;
 using Scribble.Shared.Lib.CanvasElements.Strokes;
+using Scribble.Shared.Lib.Events;
 using Scribble.Utils;
 using SkiaSharp;
