+++ Scribble/Services/CanvasStateService/Handlers/TransformReplayHandler.cs
@@ -3,6 +3,7 @@ using Scribble.Services.CanvasStateService.State;
 using Scribble.Shared.Lib;
 using Scribble.Shared.Lib.CanvasElements;
 using Scribble.Shared.Lib.CanvasElements.Strokes;
+using Scribble.Shared.Lib.Events;
 using SkiaSharp;
 
 namespace Scribble.Services.CanvasStateService.Handlers;
